Put the configuration related code to its own module
